T6 integration with T4
Changed Kafka image to Bitnami/Kafka.
Edited Docker-compose.yml file to create a topic called team6_topic on launch and clear the topic on launch if it already exists.
Added pipelines for batch data generation and streamed data generation. Updated the version of the T6 python package. Updated the readme with details on how to use new features.